Revert "split test file into 2"
This reverts commit 4ddc5ab4879fd0f524a46577021e5f9c27ee8ac2.
